Including comments from latest to Code Review. Refactoring everything to internal package and renaming to haassohnpelletstove
