--- bundles/org.openhab.binding.haassohnpelletoven/src/main/java/org/openhab/binding/haassohnpelletoven/internal/HaassohnpelletstoveConfiguration.java
@@ -1,29 +1,0 @@
-/**
- * Copyright (c) 2010-2021 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.haassohnpelletoven.internal;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-
-/**
- * The {@link HaassohnpelletstoveConfiguration} class contains fields mapping thing configuration parameters.
- *
- * @author Christian Feininger - Initial contribution
- */
-@NonNullByDefault
-public class HaassohnpelletstoveConfiguration {
-
-    public @Nullable String hostIP = null;
-    public @Nullable String hostPIN = null;
-    public int refreshRate = 30;
-}
