--- bundles/org.openhab.binding.haassohnpelletoven/src/main/resources/OH-INF/i18n/haassohnpelletoven_xx_XX.properties
@@ -1,29 +1,0 @@
-# FIXME: please substitute the xx_XX with a proper locale, ie. de_DE
-# FIXME: please do not add the file to the repo if you add or change no content
-# binding
-binding.haassohnpelletoven.name = Binding Haas & Sohn Wifi Modul
-binding.haassohnpelletoven.description = Represents a connection to a Haas & Sohn Pellet Stove Wifi Modul
-
-# thing types
-thing-type.haassohnpelletoven.oven.label = Connection to the oven
-thing-type.haassohnpelletoven.oven.description = The Connection
-
-# thing type configuration
-thing-type.config.haassohnpelletoven.oven.hostIP.label = IP Address 
-thing-type.config.haassohnpelletoven.oven.hostIP.description = Enter your IP Address of your oven in the the local network like 192.168.0.100 
-thing-type.config.haassohnpelletoven.oven.PIN.label = The oven WIFI PIN 
-thing-type.config.haassohnpelletoven.oven.PIN.description = Enter your oven WIFI PIN which you get out of the menue from the oven itself.  
-
-
-# channel types
-channel-type.haassohnpelletoven.mode.label = Mode
-channel-type.haassohnpelletoven.mode.description = Represents the actual mode of the stove
-
-channel-type.haassohnpelletoven.prg.label = Program
-channel-type.haassohnpelletoven.prg.description = Represents the program the oven is in
-
-channel-type.haassohnpelletoven.sptemp.label = Target temperature
-channel-type.haassohnpelletoven.sptemp.description = Represents the target temperature the oven should reach
-
-channel-type.haassohnpelletoven.istemp.label = Is temperature
-channel-type.haassohnpelletoven.istemp.description = Represents the is temperature the oven has right now
