+++ bundles/org.openhab.binding.haassohnpelletstove/src/main/resources/OH-INF/binding/binding.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<binding:binding id="haassohnpellestove" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns:binding="https://openhab.org/schemas/binding/v1.0.0"
+	xsi:schemaLocation="https://openhab.org/schemas/binding/v1.0.0 https://openhab.org/schemas/binding-1.0.0.xsd">
+
+	<name>Haas and Sohn Pelletstove Binding</name>
+	<description>This binding communicates with Haas and Sohn Pelletstoves through the optional
+		WLAN-Modul. It allows to power the stove on and off and receives different operation information.</description>
+
+</binding:binding>
