+++ bundles/org.openhab.binding.haassohnpelletstove/src/main/resources/OH-INF/i18n/haassohnpelletstove_xx_XX.properties
@@ -1,0 +1,29 @@
+# FIXME: please substitute the xx_XX with a proper locale, ie. de_DE
+# FIXME: please do not add the file to the repo if you add or change no content
+# binding
+binding.haassohnpellestove.name = Binding Haas & Sohn Wifi Modul
+binding.haassohnpellestove.description = Represents a connection to a Haas & Sohn Pellet Stove Wifi Modul
+
+# thing types
+thing-type.haassohnpellestove.oven.label = Connection to the oven
+thing-type.haassohnpellestove.oven.description = The Connection
+
+# thing type configuration
+thing-type.config.haassohnpellestove.oven.hostIP.label = IP Address 
+thing-type.config.haassohnpellestove.oven.hostIP.description = Enter your IP Address of your oven in the the local network like 192.168.0.100 
+thing-type.config.haassohnpellestove.oven.PIN.label = The oven WIFI PIN 
+thing-type.config.haassohnpellestove.oven.PIN.description = Enter your oven WIFI PIN which you get out of the menue from the oven itself.  
+
+
+# channel types
+channel-type.haassohnpellestove.mode.label = Mode
+channel-type.haassohnpellestove.mode.description = Represents the actual mode of the stove
+
+channel-type.haassohnpellestove.prg.label = Program
+channel-type.haassohnpellestove.prg.description = Represents the program the stove is in
+
+channel-type.haassohnpellestove.sptemp.label = Target temperature
+channel-type.haassohnpellestove.sptemp.description = Represents the target temperature the oven should reach
+
+channel-type.haassohnpellestove.istemp.label = Is temperature
+channel-type.haassohnpellestove.istemp.description = Represents the is temperature the oven has right now
